+++ broker-plugins/amqp-0-10-protocol/src/main/java/org/apache/qpid/server/protocol/v0_10/ServerConnection.java
@@ -439,7 +439,7 @@ public class ServerConnection extends ConnectionInvoker
     }
 
 
-    public long getSessionCountLimit()
+    public int getSessionCountLimit()
     {
         return getChannelMax();
     }
